Create covid-19-capture-recoveries-wikipedia.py
This code goes through the Wikipedia pages and scrapes the recoveries and other data and saves it to a CSV file. Recoveries data is only available for 6 states. However, the scrape_wiki_pages() function contains logic to identify the covid-19 data by county on any webpage and may be of reuse to scrape data from pages other than Wikipedia.