Create covid-19-wikipedia-data-scraping.py
Recoveries data for COVID-19 is limited and requires capturing data from a large number of sources. Some very generous people have been reporting the recoveries for about 6 states on Wikipedia. The Wikipedia archive pages can be used to capture the historical recovery data. This file contains python code that is intended to run in Jupyter Notebook that goes through the Wikipedia pages, scrapes the recoveries and other data, and saves it to a CSV file.