Create covid-19.ecdc-munging.py
This code was created in Jupyter Notebook to download the latest Coronavirus disease 2019 (COVID-19) data from the European Center for Disease Control (ECDC). The ECDC data is in a columnar format, so this code transforms the data to have rows with the following columns=Date, Cases, Deaths, Location (Country), GeoID, Population, and Day. The "Day" column is the number of days from when the daily number of cases # and/or deaths exceeds a limit variable. It also creates a few graphs to visualize the data on several different timelines.